move file add/remove util methods
