+++ backend/app/routers/files.py
@@ -1,9 +1,8 @@
 import io
 import json
-from datetime import datetime
-from typing import Optional, List
-
 import pika
+from datetime import datetime
+from typing import Optional, List, BinaryIO
 from bson import ObjectId
 from fastapi import (
     APIRouter,
@@ -26,6 +25,79 @@ from app.models.users import UserOut
 from app.keycloak_auth import get_user, get_current_user, get_token
 
 router = APIRouter()
+
+
+async def add_file_entry(
+    file_db: FileDB,
+    user: UserOut,
+    db: MongoClient,
+    fs: Minio,
+    file_stream: Optional[BinaryIO] = None,
+    file_obj: Optional[io.BytesIO] = None,
+):
+    """Insert FileDB object into MongoDB (makes Clowder ID), then Minio (makes version ID), then update MongoDB with
+    the version ID from Minio.
+
+    Arguments:
+        file_db: FileDB object controlling dataset and folder destination
+        file_stream: Use this when passing directly from a request e.g. UploadFile.file
+        file_obj: Use this when passing a file bytestream e.g. open(file, 'rb')
+    """
+    new_file = await db["files"].insert_one(file_db.to_mongo())
+    new_file_id = new_file.inserted_id
+
+    # Use unique ID as key for Minio and get initial version ID
+    version_id = None
+    if file_stream:
+        content_type = file_stream.content_type
+        while content := file_stream:  # async read chunk
+            response = fs.put_object(
+                settings.MINIO_BUCKET_NAME,
+                str(new_file_id),
+                io.BytesIO(content),
+                length=-1,
+                part_size=settings.MINIO_UPLOAD_CHUNK_SIZE,
+            )  # async write chunk to minio
+            version_id = response.version_id
+    else:
+        content_type = mimetypes.guess_type(file_db.name)
+        response = fs.put_object(
+            settings.MINIO_BUCKET_NAME,
+            str(new_file_id),
+            file_obj,
+            length=-1,
+            part_size=settings.MINIO_UPLOAD_CHUNK_SIZE,
+        )  # async write chunk to minio
+        version_id = response.version_id
+    content_type = content_type[0] if len(content_type) > 1 else content_type
+    bytes = len(fs.get_object(settings.MINIO_BUCKET_NAME, str(new_file_id)).data)
+    if version_id is None:
+        # TODO: This occurs in testing when minio is not running
+        version_id = 999999999
+    file_db.version_id = version_id
+    file_db.version_num = 1
+    file_db.bytes = bytes
+    file_db.content_type = content_type if type(content_type) is str else "N/A"
+    await db["files"].replace_one({"_id": ObjectId(new_file_id)}, file_db.to_mongo())
+
+    # Add FileVersion entry and update file
+    new_version = FileVersion(
+        version_id=version_id,
+        file_id=new_file_id,
+        creator=user,
+        bytes=bytes,
+        content_type=file_db.content_type
+    )
+    await db["file_versions"].insert_one(new_version.to_mongo())
+
+
+async def remove_file_entry(file_id: str, db: MongoClient, fs: Minio,):
+    """Remove FileDB object into MongoDB, Minio, and associated metadata and version information."""
+    # TODO: Deleting individual versions will require updating version_id in mongo, or deleting entire document
+    fs.remove_object(settings.MINIO_BUCKET_NAME, str(file_id))
+    await db["files"].delete_one({"_id": ObjectId(file_id)})
+    await db.metadata.delete_many({"resource.resource_id": ObjectId(file_id)})
+    await db["file_versions"].delete_many({"file_id": ObjectId(file_id)})
 
 
 @router.put("/{file_id}", response_model=FileOut)
@@ -108,19 +180,7 @@ async def delete_file(
     fs: Minio = Depends(dependencies.get_fs),
 ):
     if (file := await db["files"].find_one({"_id": ObjectId(file_id)})) is not None:
-        if (
-            dataset := await db["datasets"].find_one({"files": ObjectId(file_id)})
-        ) is not None:
-            await db["datasets"].update_one(
-                {"_id": ObjectId(dataset["id"])},
-                {"$pull": {"files": ObjectId(file_id)}},
-            )
-
-        # TODO: Deleting individual versions may require updating version_id in mongo, or deleting entire document
-        fs.remove_object(settings.MINIO_BUCKET_NAME, str(file_id))
-        await db["files"].delete_one({"_id": ObjectId(file_id)})
-        await db.metadata.delete_many({"resource.resource_id": ObjectId(file_id)})
-        await db["file_versions"].delete_many({"file_id": ObjectId(file_id)})
+        remove_file_entry(file_id, db, fs)
         return {"deleted": file_id}
     else:
         raise HTTPException(status_code=404, detail=f"File {file_id} not found")
